+++ unique_toolkit/unique_toolkit/agentic/tools/openai_builtin/code_interpreter/config.py
@@ -1,4 +1,5 @@
 from pydantic import Field
+from pydantic.json_schema import SkipJsonSchema
 
 from unique_toolkit.agentic.tools.factory import ToolFactory
 from unique_toolkit.agentic.tools.openai_builtin.base import (
@@ -14,16 +15,21 @@ Instructions:
 - All files uploaded to the chat are available in the code interpreter under the path `/mnt/data/<filename>
 - All files generated through code should be saved in the `/mnt/data` folder
 
-Instructions for displaying images and files in the chat:
+CRUCIAL Instructions for displaying images and files in the chat:
 Once files are generated in the `/mnt/data` folder you MUST reference them in the chat using markdown syntax in order to display them in the chat.
 
-- If you want to display an image, use the following syntax: `![Image Name](sandbox:/mnt/data/<filename>)`
-    - Images will be converted and shown in the chat.
-    - Do NOT display an extra download link for images a part from the markdown above.
-    - Not using markdown syntax will FAIL to show images to the user. 
-    - YOU MUST use the syntax above to display images, otherwise the image will not be displayed in the chat.
-    - For displaying a link to a file, use the following syntax: `[filename](sandbox:/mnt/data/<filename>)`
-        - Files are converted to references the user can click on to download the file
+WHENEVER you reference a generated file, you MUST use the following format:
+```
+**Descriptive Title of Graph/Chart/File** (<- linebreak is important) (You must choose a good user friendly title, Other markdown syntax such as `#` can be used too)
+[*Generating your {Graph/Chart/File}…*](sandbox:/mnt/data/<filename>)
+```
+IMPORTANT: Do NOT append a leading `!` even when displaying an image.
+Always use a line break between the title and the markdown!
+
+- Files with image file extensions are displayed directly in the chat, while other file extensions are shown as download links.
+- Not using syntax above will FAIL to show images to the user. 
+- YOU MUST use the syntax above to display images, otherwise the image will not be displayed in the chat.
+- Only the following file types are allowed to be uploaded to the platform, anything else will FAIL: PDF, DOCX, XLSX, PPTX, CSV, HTML, MD, TXT, PNG, JPG, JPEG.
 
 You MUST always use this syntax, otherwise the files will not be displayed in the chat.
 """.strip()
@@ -36,20 +42,45 @@ DEFAULT_TOOL_DESCRIPTION_FOR_USER_PROMPT = ""
 
 
 class OpenAICodeInterpreterConfig(BaseToolConfig):
-    upload_files_in_chat: bool = Field(default=True)
-
-    tool_description: str = DEFAULT_TOOL_DESCRIPTION
-    tool_description_for_system_prompt: str = DEFAULT_TOOL_DESCRIPTION_FOR_SYSTEM_PROMPT
-    tool_format_information_for_system_prompt: str = (
-        DEFAULT_TOOL_FORMAT_INFORMATION_FOR_SYSTEM_PROMPT
-    )
-    tool_description_for_user_prompt: str = DEFAULT_TOOL_DESCRIPTION_FOR_USER_PROMPT
-    tool_format_information_for_user_prompt: str = (
-        DEFAULT_TOOL_FORMAT_INFORMATION_FOR_USER_PROMPT
+    upload_files_in_chat_to_container: bool = Field(
+        default=True,
+        description="If set, the files uploaded to the chat will be uploaded to the container where code is executed.",
     )
 
-    expires_after_minutes: int = 20
-    use_auto_container: bool = False
+    tool_description: str = Field(
+        default=DEFAULT_TOOL_DESCRIPTION,
+        description="The description of the tool that will be sent to the model.",
+    )
+    tool_description_for_system_prompt: str = Field(
+        default=DEFAULT_TOOL_DESCRIPTION_FOR_SYSTEM_PROMPT,
+        description="The description of the tool that will be included in the system prompt.",
+    )
+    tool_format_information_for_system_prompt: SkipJsonSchema[str] = Field(
+        # Since the tool is executed in Azure, it's not always possible to have dynamic format information
+        default=DEFAULT_TOOL_FORMAT_INFORMATION_FOR_SYSTEM_PROMPT,
+        description="The format information of the tool that will be included in the system prompt.",
+    )
+    tool_description_for_user_prompt: SkipJsonSchema[str] = (
+        Field(  # At the moment, this is not appended to the user prompt
+            default=DEFAULT_TOOL_DESCRIPTION_FOR_USER_PROMPT,
+            description="The description of the tool that will be included in the user prompt.",
+        )
+    )
+    tool_format_information_for_user_prompt: SkipJsonSchema[str] = (
+        Field(  # At the moment, this is not appended to the user prompt
+            default=DEFAULT_TOOL_FORMAT_INFORMATION_FOR_USER_PROMPT,
+            description="The format information of the tool that will be included in the user prompt.",
+        )
+    )
+
+    expires_after_minutes: int = Field(
+        default=20,
+        description="The number of minutes after which the container will be deleted.",
+    )
+    use_auto_container: bool = Field(
+        default=False,
+        description="If set, use the `auto` container setting from OpenAI. Note that this will recreate the container on each call.",
+    )
 
 
 ToolFactory.register_tool_config(