+++ unique_sdk/CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.10.43] - 2025-11-14
-- Add get_groups, delete and update functions.
+- Add get, delete and update groups functions.
 
 ## [0.10.42] - 2025-11-14
 - Add get_users function.