+++ unique_sdk/CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.10.44] - 2025-11-14
+- Create group and manage users functions.
+
 ## [0.10.43] - 2025-11-14
 - Add get, delete and update groups functions.